# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI App Builder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  description: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
  fileCount: number;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  projectId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface ProjectFile {
  id: number;
  projectId: number;
  path: string;
  content: string;
  language: string;
  createdAt: string;
  updatedAt: string;
}

export interface FileInput {
  content: string;
  language?: string;
}