# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI App Builder API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Project management
  - name: messages
    description: Chat messages
  - name: files
    description: Project files

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/recent:
    get:
      operationId: getRecentProjects
      tags: [projects]
      summary: Get recently updated projects
      responses:
        "200":
          description: Recent projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{projectId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List chat messages for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message and get AI response
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: AI response message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /projects/{projectId}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List all files in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectFile"

  /projects/{projectId}/files/{filePath}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get file content
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
        "404":
          description: Not found
    put:
      operationId: updateFile
      tags: [files]
      summary: Create or update a file
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileInput"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Project:
      type: object
      required: [id, name, description, createdAt, updatedAt, messageCount, fileCount]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        messageCount:
          type: integer
        fileCount:
          type: integer

    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string

    Message:
      type: object
      required: [id, projectId, role, content, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1

    ProjectFile:
      type: object
      required: [id, projectId, path, content, language, createdAt, updatedAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        path:
          type: string
        content:
          type: string
        language:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    FileInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        language:
          type: string