# lib/api-zod/src/generated/types/projectFile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI App Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectFile {
  id: number;
  projectId: number;
  path: string;
  content: string;
  language: string;
  createdAt: Date;
  updatedAt: Date;
}